+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 0.1.5 - xxxxxxxx
 
-- Modified: `paramvar` in _quants_
+- Modified: `paramvar`,`histc`
 
 ### 0.1.4 - 20150309